+++ ufo/automator/ui_control/screenshot.py
@@ -5,12 +5,16 @@ from typing import List
 
 from PIL import Image, ImageDraw, ImageFont, ImageGrab
 from pywinauto.win32structures import RECT
-from typing import Dict
+from typing import Dict, Optional
 from abc import ABC, abstractmethod  
+from io import BytesIO
 
 from ...config.config import Config
 from ...utils import number_to_letter
 
+import base64
+import mimetypes
+import os
 import win32api
 import win32con
 import win32gui
@@ -363,6 +367,41 @@ class PhotographerFacade:
         result.save(output_path)
         
         return result
+    
+    
+    @staticmethod
+    def image_to_base64(image: Image):
+        """
+        Convert image to base64 string.
+
+        :param image: The image to convert.
+        :return: The base64 string.
+        """
+        buffered = BytesIO()
+        image.save(buffered, format="PNG")
+        return base64.b64encode(buffered.getvalue()).decode("utf-8")
+
+    
+    @staticmethod
+    def encode_image_from_path(image_path: str, mime_type: Optional[str] = None) -> str:
+        """
+        Encode an image file to base64 string.
+        :param image_path: The path of the image file.
+        :param mime_type: The mime type of the image.
+        :return: The base64 string.
+        """
+        
+        file_name = os.path.basename(image_path)
+        mime_type = mime_type if mime_type is not None else mimetypes.guess_type(file_name)[0]
+        with open(image_path, "rb") as image_file:
+            encoded_image = base64.b64encode(image_file.read()).decode('ascii')
+            
+        if mime_type is None or not mime_type.startswith("image/"):
+            print("Warning: mime_type is not specified or not an image mime type. Defaulting to png.")
+            mime_type = "image/png"
+            
+        image_url = f"data:{mime_type};base64," + encoded_image
+        return image_url
